refactor: split api/types.rs into type defs and mapping module
Move `From` impls and `convert_to_events` out of api/types.rs into a
new api/mapping.rs, leaving types.rs with UniFFI type definitions only.
Tests for `convert_to_events` move with the function.
